Move tooltip up to AlignmentView.
